backend/requirements/base.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
Pillow==10.0.1
django-extensions==3.2.3
celery==5.3.4
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
elasticsearch-dsl==8.11.0
drf-yasg==1.21.7
drf-spectacular==0.26.5
psutil==5.9.6
requests==2.31.0